+++ utils/preval/access-token.preval.ts
@@ -1,5 +1,4 @@
-import axios from "axios";
 import preval from "next-plugin-preval";
-import { fetchAccessToken } from "./functions";
+import { fetchAccessToken } from "../functions";
 
 export default preval(fetchAccessToken());
